Fix type_structure_classname crash
Summary:
HHBBC will convert type_structure($foo, $str)['classname'] into a call
to type_structure_classname($foo, $str) instead, which can be
optimized specially in the JIT. However, it does so without knowing
anything about the inputs, so there's no guarantee that 'classname' is
actually present in the resolved type-structure! The implementation of
type_structure_classname uses asCStrRef(), which will assert if the
value is Uninit (not present). Instead use the proper getter which
throws if the value isn't present. We still assume the value is a
string if present because that's a type-structure invariant and we
assert that all over the place.

Differential Revision: D30489637

fbshipit-source-id: 9310b634d00627b19e70cf3d5d24837c51201c94