Annotate contexts for Exception
Summary:
This mirrors the contexts in the hhi, with some tweaks where it makes sense.

The main things missing now are `getMessage` and `toString` (which calls `getMessage`). Unfortunately, `getMessage` is often overriden in ways that are impure, so we need some kind of polymorphic solution for that.

`setPrevious` and `__prependTrace` probably ought to be `write_this_props` but the typechecker doesn't support it so `write_props` will do for now.

Reviewed By: oulgen

Differential Revision: D30261108

fbshipit-source-id: adcdff784a036ecc224adc3e0a20b8e26b6eb2e2