Autoloading is not const-foldable in sandbox mode
Summary:
This method is marked foldable, but uses autoloading internally. In sandbox mode, when we hit retranslate-all, we can const-fold an incorrect "false" value for this method because autoloading is blocked during irgen-builtin evaluation.

If we care to bring back this optimization in repo mode, we can do so, but I doubt it matters. This helper is only used on slow reflection paths.

Reviewed By: ricklavoie

Differential Revision: D30522582

fbshipit-source-id: 1fb9c06adad1471d30a0bd93721a7a6259b8054c