Add an `indexedMethodAttributes` option to the SymbolMap constructor
Summary:
Hack allows developers to decorate files, classes, and methods with attributes, like so:

```
<?hh

<<file: SomeFileAttr>

<<SomeClassAttr>
class SomeClass {
  <<SomeMethodAttr>>
  public function someMethod(): void {}
}
```

Facts allows developers to query for all classes/files/methods using a given attribute, which allows for lots of neat use-cases.

Our codebase has *tons* of method attributes, to the point where most of our DB is devoted to recording them. But right now we only need to index a handful of method attributes.

In the long run, I want Facts to scale so we can index all the things. But in the short run, it's silly to spend so much time and space indexing data that nobody cares about. So I'm adding a new `indexedMethodAttributes` list. If the list exists, then we'll only record method attributes which are in the list.

Reviewed By: bdchik

Differential Revision: D30378148

fbshipit-source-id: 98afcc847cd0bd4d9d591b8274e34455a3ad0e6e