Bubble method attribute allowlist to an .hhvmconfig.hdf option
Summary:
This diff continues the work begun in D30378148, creating a new option in `.hhvmconfig.hdf` called `Autoload.IndexedMethodAttributes`. If `.hhvmconfig.hdf` has this option, we'll only record the method attributes in this list.

Note that we include all `.hhvmconfig.hdf` options in the hash used to store the `FactsStore` in memory, and the hash used in the DB's filename.

Reviewed By: bdchik

Differential Revision: D30378149

fbshipit-source-id: a87c6343c56894b280b7ecc37dca76b84a586198