Fix regression of reified static method calls
Summary:
In a previous change, we had regressed the error checking of static method calls on reified generics in Expression Trees. Namely, the following example:

```
function takes_reified<reify T as Foo>(): void {
  Code`T::bar()`;
}
```

This is because the previous change had extracted/lifted the function pointers of the static methods out of the body of the virtualized expression and assigned them to temporary variables as part of `et_splices`. As a result, our check within `reified_check.ml`

```
      | ( _,
          pos,
          FunctionPointer (FP_class_const ((ty, _, CI (_, class_id)), _), _) )
        when Env.is_in_expr_tree env ->
        let (_env, ty) = Env.expand_type env ty in
        begin
          match get_node ty with
          (* If we get Tgeneric here, the underlying type was reified *)
          | Tgeneric (ci, _) when String.equal ci class_id ->
            Errors.reified_static_method_in_expr_tree pos
          | _ -> ()
        end
```

No longer traverses the function pointer node as `et_splices` is not considered part of `in_expr_tree` in typing. Instead, it sees just the variable substitution, at which point, it does not throw an error because that is not the node being matched.

This diff traverses the function pointer nodes as part of `in_expr_tree`. Since the previous diff assigned the toplevel function pointers to temporary variables as part of `et_splices` and we do not want to think that `et_splices` are part of the `in_expr_tree` in typing, we split off the temporary assignments from the splices and store them as  `et_function_pointers`. Then we once again traverse `et_function_pointers` as part of `in_expr_tree`.

Thus, previously, for the following example:

```
Code`${ $x } + foo() + Foo::bar()`;
```

We would have created the following split:

```
et_splices = {
  $0tmp1 = $x;
  $0fp1 = foo<>;
  $0sm1 = Foo::bar<>;
}

et_virtualized_expr = {
  ${ $x } + Code::symbolType($0fp1)() + Code::symbolType($0sm1)()
}
```

Now we create the following split:

```
et_splices = {
  $0tmp1 = $x;
}

et_function_pointers = {
  $0fp1 = foo<>;
  $0sm1 = Foo::bar<>;
}

et_virtualized_expr = {
  ${ $x } + Code::symbolType($0fp1)() + Code::symbolType($0sm1)()
}

```

Reviewed By: Wilfred

Differential Revision: D30292855

fbshipit-source-id: 212ac9298b9d0477d4740871151840fcad363c20