add Locale\ (#8871)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/8871

As part of string purity, we need to remove the implicit use of the thread locale.

To unblock that, we need to be able to explicitly retrieve and set it.

This diff is just about `Locale\`: the locale-aware string functions used here are
meant to be fairly close to libc, and are definitely not the desired API for `Str\`,
which will be a separate PR/diff.

This is meant to be enough `Locale\` for practical use of `Str\` and setting the request locale. It is not intended to be complete - in particular, it should probably be possible to inspect a locale without going via `\setlocale()` or depending on the debuginfo. Leaving that for later for use cases to develop.

Reviewed By: viratyosin

Differential Revision: D29311121

fbshipit-source-id: 79845755f09a3f8c11a4f81fe694e003bb416bc8