Remove members_fully_known field
Summary: For both legacy and shallow decls we maintain a field `members_fully_known` that records for each class whether all the ancestors of the class are "known". Once upon a time, we legitimately extended "unknown" PHP classes, I think, but now the only use of this field is to prevent a few downstream errors when a class is not found. As you can see from the test output, it doesn't make much difference. Let's kill it.

Reviewed By: CatherineGasnier

Differential Revision: D30511168

fbshipit-source-id: b157941eedf6f617111acd8ac945c0f21eaba9e9