UNSAFESingletonMemoizeParam
Summary:
This diff introduces a new marker interface called UNSAFESingletonMemoizeParam, which promises Hack that the class in question is a singleton per instance key. We do not enforce that the parameters are actually singleton per instance key. From a runtime perspective, this interface does nothing.

Note: please let me know if I put this in the correct hhi/systemlib places, or if something harder to discover may be better. We don't have many use cases in mind for this(mainly just a single one), so its okay for it to be rather undiscoverable, since it is unsafe.

Reviewed By: viratyosin

Differential Revision: D30519650

fbshipit-source-id: 9e3cfee6eb059fb1b10243c3f0514f19fff5c50f