Loop analysis
Summary:
This diff adds loop fixed point analysis to readonly check. The algorithm works as follows:

- Given a body of a loop, we need to evaluate the readonlyness of the body up to X times, where X is the number of mutable values in the loop (including new ones created during the loop's execution, of which there's a finite number).

Invariant: the number of readonly values in an environment cannot ever decrease.

Each iteration of the loop, one of two things can happen:
- The local environment's number of readonly values does not change
- The local environment's number of readonly values increases

In the former case, we can stop looping, since we are at a fixed point: the number of readonly values cannot increase (nor can it decrease) after an iteration, so we are done.
In the latter case, the number of readonly values increases, so we loop again.

Since the number of readonly values must always increase for each iteration of the loop, and the maximum number of readonly values in the environment is the number of local variables, we know this loop will terminate after X iterations.

This can be slow with nested loops, so we may need to add caching or other mechanisms for figuring out loop fixed points if its ever too complicated or slow. But we will deal with perf in a different diff as needed.

Reviewed By: oulgen

Differential Revision: D30344623

fbshipit-source-id: 9580519fc434c546c7d2917bd8a5f85fba819ee9