Fix markers for inlined opt builtins
Summary:
Now that we have a proper inline adjustment pass, lying about the SrcKey here is unnecessary (and just wrong). The fixupSk will have the proper adjustments.

Prior to this diff, anything after the first mayThrow instruction in an inlined builtin will be incorrectly attributed to the bytecode instruction following the builtin call.

Reviewed By: ricklavoie

Differential Revision: D30628339

fbshipit-source-id: 1aa8713c62425a307f483a37cb8d17b63ec1db48