Ban readonly static methods and variadic arguments
Summary:
This diff adds two checks that ban readonly on static methods, and readonly on variadic arguments.

Static methods don't need to be marked readonly because calling them cannot result in modifying any instance properties. (If you want a static property to be readonly, you can mark that as such).

Variadic arguments could be supported in the future (they would be passed as a readonly vec of arguments), but we don't plan on it for the time being.

Reviewed By: oulgen

Differential Revision: D30648140

fbshipit-source-id: 59e1608385e3cc1f461d87be3696096cbdc272a0