Use direct decl parser to update naming table
Summary:
We should use the direct decl parser results for the naming phase.

The direct decl parser behaviour is different from the legacy parser: parser errors in the body might be ignored! As such, if we use the legacy parser result to update the naming table, we might erroneously remove entries! When typechecking files using that entry, erroneous "Name unbound" errors will be reported. Those are unrecoverable, because decl diffing won't catch any changes (the direct decl parser returns the same result with or without parsing errors), and no fanout will be computed.

Differential Revision: D30332172

fbshipit-source-id: d0cc8de0d460596b0f07b7bbdc986cedbdd23c00