Build hashbrown with nightly flag
Summary:
Build hashbrown with the nightly flag enabled.

As part of a Hack project, we want to use hashbrown's HashMap with a custom allocator (short story: we want to have the HashMap live and allocate in shared memory).

However, to get access to this functionality, we must build the hashbrown crate with the nightly flag.

Is this dangerous?I [asked in **Rust at Facebook**](https://fb.workplace.com/groups/rust.language/posts/6375156715866223): the flag appears to be additive, so the consensus is that it's fine to enable this FB-wide

Reviewed By: mjhostet

Differential Revision: D30070281

fbshipit-source-id: f165ac1beaf5a44d5cb5b9012aebad5dc54beb95