Relax queries if the repo violates the one-definition-rule
Summary:
In Hack, it's an error to define the same class in more than one file. If you try to autoload that class, your code will not work. But it's not exactly clear what "your code will not work" should mean.

In the past, we've chosen to treat duplicate-defined classes as nonexistent, so that defining a class in more than one file is the same as leaving the class completely undefined. We still have to store facts for the duplicate-defined class, because you might remove all but one of the duplicate definitions. But we'll strip duplicate-defined classes from results, and we'll refuse to autoload any symbols defined in more than one file.

There's a pretty hefty performance penalty associated with stripping duplicate-defined classes from results, and the Hack typechecker already provides a layer of protection against duplicate-defined symbols. It may be safe to relax our contract here, and begin returning nondeterministic answers when a symbol has more than one definition.

This diff removes all the code which treats ODR violations as undefined symbols, and rips out all the test code which requires that we treat them in this way.

Reviewed By: bdchik

Differential Revision: D30561239

fbshipit-source-id: 4ef165df1ca6f46e12a4721b9d9a520aad245cde