don't compute fan in deps when doing typechecking
Summary: Now that we are pivoting to old decls for fanout reduction, we don't need this. I've also found anecdotally that this path does sometime cause the command to fail. eg. on the www commit 76af3f239cffdfa3dd5c3aedbe3916b0601969cf. This also means we no longer need to do a full init when running this command, which should help with capacity/perf.

Differential Revision: D30734255

fbshipit-source-id: bf505362a77a564b6120a47d9e38318f0e7545d3