Remove the Record type
Summary:
It's never been used and it's not fully complete. There's no plans to work on it soon. Rather than letting the code sit and bit-rot, remove it. If we reintroduce it, we can start clean with whatever the final semantics we want are.

I'm only touching the runtime because I don't know enough about the parser side to remove it.

Reviewed By: oulgen

Differential Revision: D30723023

fbshipit-source-id: 095e3f968faaa19b4501e1e8baa35e4ce149de55