Improve type-structure pre-resolution logic in HHBBC
Summary:
HHBBC has some support for pre-resolving a type-structure ahead of
time. Improve it greatly. It can now handles almost every case the
runtime can. Going forward, we must ensure that the runtime and HHBBC
agree exactly on how resolution works.

Whereas before HHBBC would only pre-resolve "anonymous"
type-structures (those in the bytecode). It can now pre-resolve
type-aliases and type-constants as well. Since type-constants can use
this:: and therefore be resolved differently in different subclasses,
HHBBC will copy the type-constant into every subclass and resolve it
separately there. As a result, this:: is always known and we can
(almost always) pre-resolve the type-constant. The inference enum is
also set at this point.

Add a helper function to the Index, called
lookup_class_type_constant. It behaves like lookup_class_constant, but
is responsible for type-constants. Change over the logic for handling
type_structure and type_structure_nothrow over to use that
instead. The override bit on constants is now unused and is removed.

Differential Revision: D30637487

fbshipit-source-id: 21f2ba82707bfdb7260d167e34640ae7750ad9db