Type-structure is/as checks should invoke autoloader
Summary:
The code to match a type-structure against a TypedValue for the
purpose of is/as checks only called Class::lookup when it needed to
map a string to a Class. This mainly works because type-structure
resolution uses Class::load during resolution, thus any class
mentioned in a type-structure should already be loaded.

Where this falls apart is if we're pre-resolved the type-structure. If
we pre-resolved it, the type-structure may refer to classes which are
guaranteed to exist, but may not be actually loaded yet. Use
Class::load instead of Class::lookup to deal with this. The
autoloading should never fall, as a resolved type-structure cannot
refer to a non-existent class. This all becomes a much bigger problem
once HHBBC starts more aggressively pre-resolving type-structures.

Differential Revision: D30636740

fbshipit-source-id: b7fc917b25eb381741e1ef4e3a879bd681419777