Improve class constant analysis in HHBBC
Summary:
Improve class constant analysis in HHBBC. HHBBC wasn't really dealing
with constants defined by 86cinits at all. If we could (eventually)
resolve them a scalar value, we'd use the type, but not
otherwise. There's some constants which can never be scalars (like
objects). Instead, store the type that is returned from the 86cinit in
the Index. When looking up a class constant, if a scalar value isn't
known, you can get the best known type from the Index. Apply all the
usual rules about refinements.

Add a new function to the Index, called lookup_class_constant. This is
meant to be the single source of truth for class constants during
analysis. It is similar to lookup_static and other similar
functions. It is given a class and name, and will find the best known
type for that constant, along with other metadata. It does this by
unioning together the results for all possible subclasses (including
using the above mentioned 86cinit data). lookup_class_constant is only
responsible for "normal" constants (non-type, non-context).

Change the implementation of the ClsCns* bytecodes to use
lookup_class_constant, as well as the few other places that was
looking up class constant information.

Differential Revision: D30635081

fbshipit-source-id: 83983ebfc58c2a2385b2d5948d1ba0ffaece84cc