In HHBBC, deal with calls to unknown functions which we know will always throw
Summary:
In HHBBC, in fcallKnownImpl, we deal with the case where we
know the call will always throw. Either the function body itself
always throws, or the call itself throws (for example, if the $this is
null). In either case, we properly mark the bytecode as unreachable
and return Bottom.

We don't do any of this in the fcallUnknownImpl case. The idea here is
that we don't know the function we're calling, so we can't possibly
know it will always throw or not. This is missing one case however. If
the call is a nullthrows, and $this is known to be null, we know the
call will throw, even if we don't know the destination. In this case,
we actually pessimize things, pushing an InitCell.

This can cause a monotonicity violation, since refining the type of
$this can cause the output type to expand.

Fix this by detecting this case, and marking the bytecode as
unreachable, as in the known function case.

Reviewed By: jano, mofarrell

Differential Revision: D30490390

fbshipit-source-id: 649c57361352fe2d40a51b6ad674c40ca29f4a73