Leverage type-structure pre-resolution more in the JIT
Summary:
Take advantage of HHBBC's pre-resolution of type-structures more in
the JIT, namely for optimizing HH\type_structure and family. The
existing JIT logic was a bit suboptimal. It was actually doing more
checks than necessary. It's easy to ensure that a type-constant will
never have a value with it's lower bit set, unless it's resolved (so
abstract type-constants would fail). Therefore it suffices to merely
check the bit. If it's set, clear it, and take the fast
path. Otherwise take the slow path.

To make this check even more optimal, use a new vasm instruction
"btr", which tests a bit, clears it, and returns the bit's value in a
flag. This lets us avoid doing an andqi, then a xorqi.

We can remove the check entirely by using the invariance enum. If the
invariance indicates the type-structure does not change, we can simply
const prop it. Otherwise, we can assume it will always be present, and
thus do not need to check the bit at all. Similar optimizations can be
done with classname and loading the cached classname field.

Differential Revision: D30637548

fbshipit-source-id: c05b8f027f69147cc4b362f3c74aca195a8ae8c7