Fix nullsafe call operator with inout arguments
Summary:
The combination of the nullsafe call operator with inout arguments was
very broken. In the null case, we only pushed a single null onto the
stack, despite the subsequent bytecodes expecting multiple outputs
pushed. As a result, we'd crash. Both the interpreter and JIT had this
issue. I'm not sure what the actual behavior should be here, but I
made them just push the inout arguments onto the stack unchanged
(which makes sense, since the call never happened).

HHBBC also had issues. It ignored the null case completely, for the
most part. Therefore it could decide a call would always throw, when
in reality it might not because of the null case. HHBBC also failed to
union in the inout argument types to the outputs, meaning it would
infer the incorrect types. Fix all of this.

Differential Revision: D30634708

fbshipit-source-id: 4d9ca43c6d85221d5102b0d3496698b2ba47576b