Create Two Log Handlers: Cronolog and the HHVM Error Log
Summary: Logging in Facts is fairly invisible to users and can only be seen by turning on trace flags.  This implements two log handlers using the folly Logging system.  One log handler wraps the Cronolog class used in HHVM.  This will create a log file dedicated to Facts and the Cronolog class will take care of rotating the file on a daily basis.  The other log handler simply calls Logger::Error() to include information in the regular HHVM log as well as the configured thread log.

Reviewed By: jthemphill

Differential Revision: D30567044

fbshipit-source-id: 16909bc1c5fbc7c3c9505e5cf2a5f9f4f5aa4ee3