Asynchronous Logwriter
Summary: Logging is great and all, but logging within the critical path may have a performance impact, especially if flushing after each write.  This creates a logwriter wrapper which makes the underlying logger asynchronous by changing the logging call into a queue operation and managing the actually writing to disk on a different thread.

Reviewed By: jthemphill

Differential Revision: D30567072

fbshipit-source-id: b57a935418e925b2b844e9ea33c038664ef11df7