Support UPDATE=always in review.sh
Summary:
Add a flag that allows users to update all the `.exp` files. This is particularly useful for changing NAST or TAST .exp files when a new AAST field is added.

Update verify.py to use this new `UPDATE` flag, and to suggest the `always` when there are a large number of changes.

Reviewed By: vsiles

Differential Revision: D30716090

fbshipit-source-id: 3054d435136bfe95164eeb7ffa91c0cc9e3020f4