Get rid of profile code area, use main area instead
Summary:
The code.prof section was used to keep the main portion of the profile
translations.  Getting rid of this area frees up scarce memory address space
below the 1GB virtual address.

The original motivation to put code in a separate area instead of in main was
to improve the code locality of the live and optimized translations that are
normally put in the main code area.  These days, the profile code is collected
in a separate HHVM invocation in the jumpstart seeders, so that motivation
doesn't hold anymore.  And, even when jumpstart is not used, the vast majority
of the cycles are normally spent in optimized translations, which are generated
in one shot via retranslate-all and placed in a separate code area anyway
(code.hot).

This diff keeps the Eval.JitAProfSize, and the main area is allocated with
Eval.JitASize + Eval.JitAProfSize. This will be cleaned up later.

Reviewed By: ricklavoie

Differential Revision: D30518528

fbshipit-source-id: a71b9f3ba98b08c23224da3820a24653c6be9054