Get rid of hot code area, use main area instead
Summary:
The original motivations for putting the optimized code in a separate hot code
area, right before the main code area, was to be able to pack it all together
and map to huge pages.  With jumpstart, this already happens naturally, since
the optimized code is JITed before any live translation.  In non-jumpstart mode,
the retranslate-all mechanism already makes sure that all the optimized code is
produced consecutively in the translation cache, but we'd still miss on the
ability to map all the optimized code onto huge pages, because not all the main
code area is normally mapped to huge pages.  To address that, this diff forces
mapping the entire main code area onto huge pages when jumpstart is not used,
trading a small increase in RSS for better performance in this case.

Reviewed By: ricklavoie

Differential Revision: D30518526

fbshipit-source-id: cd6ca8f5d6aea150d934972bc44aff5a04b6a8f2