Make 0 sampled coeffect experiments better
Summary:
When we run experiments to determine cost of coeffect enforcement - cost of raising warnings, we want to use `-vEval.CoeffectViolationWarningSampleRate=0` so that the warnings are not raised; however, this still pays the cost of entering the C++ and figuring out the error message etc.
On the other hand, if I were to not emit `RaiseCoeffectsCallViolation`, instruction when the sample rate is 0, then dce and load/store elim would kill the instructions for coeffect enforcement.
This solution ensures that coeffect enforcement is as expensive as it is when it always succeeds after properly checking.

Reviewed By: jamesjwu

Differential Revision: D30769802

fbshipit-source-id: f6794a0c1325d504a20d278b619393a534099833