Provide a .mli file for SharedMem.ml
Summary:
Provide an *.mli file for the `SharedMem` module.

The "design" of our `SharedMem` module is hard to grasp. An *.mli file should improve that. It should also allow me to come up with a more principle design.

I'd like to have a uniform cache type. This will help me adapting SharedMem to support an evictable backend

 ---

In addition to providing the mli file, I've introduced some minor changes:

- I've introduced some submodules, grouping some functions according to domain. I've introduced a:
    1. GC module to contain GC-specific functions
    2. SMTelemetry module to contain telemetry-specific functions (SMTelemetry not to clash with the widely used Telemetry module)
    3. RawAccess module that provides raw access to the underlying shared memory table
    4. DepTable module that groups together dep table functions (these will soon (tm) be gone)
- I've changed some debtable functions to accept named arguments, because previously they just took two unnamed string arguments.
- Some comments have been moved to the mli file. Some comments have been reworded. Some comments have been removed because they weren't telling the truth,

Reviewed By: Wilfred

Differential Revision: D30789148

fbshipit-source-id: 8040569d128c2c8edb7855619ea6ee414b5d88f9