Add Str\foo_l UTF8-capable and locale-aware functions. (#8882)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/8882

We have a single object that represents encoding + locale, similar to the `en_US.UTF-8` strings being used with `setlocale()`.

For now, we only support single-byte and UTF-8, as we require that all pairs are potentially valid thread/process locales - UTF-16 isn't because C locales can't have null bytes except as terminators. We could (and probably should) support UTF-16 in the future, but we would have to make Locale\set_native() throw an exception when a UTF-16 is passed.

As for design:
- if `Str\foo()` exists, so does `Str\foo_l()`, and `Str\foo_l()` will always take a `Locale\Locale` object as the first parameter.
- in most cases, the rest of the parameters match; in some cases, the locale defines previously optional parameters, so I've removed the parameter in the `_l` version

`_Private\_Str`: not a supported API. That said, it follows the BSD libc conventions; most glibc _l functions also exist in BSD libc, and BSD libc has many more.

I went with that convention for `_Str` mostly to unblock implementation. I've not used it for the public API as:
- the parameter convention is 'the last non-variadic non-format-string parameter'; this isn't super clear and feels inconsistent, but makes sense for C. Given that past work shows edge cases like `vsprintf_l(char* str, locale_t loc, const char* format, va_list ap)`, we can avoid them with a more consistent design now.
- unintuitive naming: the case-sensitive version of `strcasecmp()` is `strcoll()`, not `strcmp()`; `strcasecmp()`  and `strcoll()` are locale-sensitive, `strcmp()` isn't.
- having `foo` and `foo_l` with similar args may raise awareness of the need for locale-sensitivity

Reviewed By: oulgen

Differential Revision: D30458827

fbshipit-source-id: 75159aec5757152b504d449acd0adcb4f338c3c4