Mark invariant_violation as pure
Summary:
The callback only ever gets set once at the beginning of the request for FB WWW. This means that this global read is technically pure within the request.

There are alternative proposed solutions on the task but none of them are viable as making configs read only would require significant work. The ideal solution would be moving the callback inside the runtime but that would also be a lot of work.

For the time being, lie and punt the work.

Reviewed By: rodmk, viratyosin

Differential Revision: D30846138

fbshipit-source-id: 71ec45cbc143eac903c4f4ec65aebf8f5892fc24