Changes to run on external machine (#8879)
Summary:
Made a few changes in the code, so that it can be run on the external machine. To be compatible with both internal and external ways of running. We need to find a way to detect the code is running internal environment or external.

We can start a discussion on how to detect in this P/R thread, I will adjust my code later.

Pull Request resolved: https://github.com/facebook/hhvm/pull/8879

Test Plan: Existing tests pass  both external and internally

Reviewed By: zhyty

Differential Revision: D30813937

Pulled By: jamesjwu

fbshipit-source-id: a31aa88ff0876a8a477ddaee9ea0fbb8ebaa7f82