Add option to control default locale of `Str\` functions
Summary:
Aim is to move from 0-3 relatively quickly. The error option (2) is useful to fail at the first call
instead of at some later point.

I also changed the implementation of `HSLLocale` as it needs to be copyable to be nativedata, and it wasn't with a
raw ops pointer. This was pre-existing, but attempting to cache locales made this very evident.

Reviewed By: oulgen

Differential Revision: D30847300

fbshipit-source-id: b241f328dbdce0b1224aea445bb2a8a2d8c359bc