Readonly lambdas
Summary:
This diff handles readonly lambdas and closures in general. When entering a lambda, if the lambda is readonly_this, all of the variables captured by the closure are readonly.

We handle this properly in visit_fun_ vs. visit_fun_def. Note that visit_fun_def calls visit_fun_ so the parameters get set by it correctly.

Reviewed By: oulgen

Differential Revision: D30714694

fbshipit-source-id: e15906c928647265567efeb7385c73d87dcb909e