Fix deadlock in Treadmill crash reporting
Summary:
When dumping the Treadmill from crash-reporter, we attempt to
grab the lock (but continue anyways if we can't). However, we never
attempt to drop the lock when we're done. Thus the Treadmill remains
locked and any later attempt to insert into it will deadlock. Namely,
this can happen during PHP stacktrace reporting (which happens after
the Treadmill dump). Always drop the lock, even in crash mode.

Also make the treadmill output a little prettier.

Reviewed By: oulgen

Differential Revision: D30888142

fbshipit-source-id: 565d758984de81a75c959fdd378576594cfe21c6