Fast exit from subtyping with Terr
Summary: The deferred decl mechanism works as follows: when the number of decl queries reaches some limit, all subsequent queries to `get_class` will return `None`. Inside the type-checker, this often manifests as type `Terr`. So the somewhat-ill-defined processing of `Terr` is crucial to the performance of the checking of code downstream of the point when deferred decls kick in. One particularly bad issue (it seems) is with large intersections produced by repeated type refinement (the `as` or `is` constructs). Instead of decomposing subtype checks `err <: t1 & ... & tn` into `err <: t1` ... `err <: tn` it is quicker to exit early with success.

Differential Revision: D30899770

fbshipit-source-id: e183b5419912fba3c47b928b3e5320892d347401