Add AsyncPool::connectAndQuery to HHVM extension
Summary:
**Background**

We'd like to expand connection pooling to more use cases in WWW. With jupyung's recent changes to connection pooling, we believe that we might even be able to experiment with broad rollouts (like rolling out for all of async).

However, rolling out connection pooling today requires a bit of work. The vast majority of WWW MySQL queries use the connectAndQuery API (via DBQuery), but connection pooling is currently only supported for generating individual connections (via AsyncDBConn::genConn). This means that for any DBQuery callsite where we'd like pooling, we need to first convert it to a DBQueryFromConnection. Not only does this require some eng work, converting DBQuery to DBQueryFromConnection exposes the code to an increased risk of timeouts.

----

**This Diff**
Add a `connectAndQuery` API to `AsyncMysqlConnectionPool` in the MySQL HHVM extension.

**Next Steps**
Once this is available in WWW, we'll modify the API for DBQuery to allow pooled connection (similar to [what we do for AsyncDBConn today](https://fburl.com/code/8cz2v7er)).
- With that in place, we'll look at doing broad rollouts for connection pooling in WWW.
- We'll also start converting high-potential DBQuery callsites to use pooling without changing them to DBQueryFromConnection

Reviewed By: jupyung

Differential Revision: D30717914

fbshipit-source-id: 108df5ec18bfff760b6908c9f6f170fe53b6cb38