Fix dynamic_meth_caller coeffect enforcement
Summary:
`HH\dynamic_meth_caller('C', 'foo')` ends up creating a lambda (which is a class with an invoke function). So the question is what should the coeffects of this lambda be. If the caller of this lambda can call `(new C)->foo` (i.e. it has necessary context), this lambda should be able to call it too, similarly, the caller should be able to call this lambda always, so the lambda needs identical coeffects to its caller which ends up being tricky since the coeffects change based on the call site not definition site.

The solution is the same thing we do for generator methods where we say the coeffects of this method is that of its caller, so we stash the coeffects of its caller in a local on the method so that once we call anything we can compare the coeffects of the caller with the next callee, so the middle function acts as a transferring spot. This is simpler for generator methods than lambdas since lambdas would need to now support two different ways to get coeffects (1) inherit from construction site (2) inherit from caller.

Reviewed By: jamesjwu

Differential Revision: D30905002

fbshipit-source-id: 523adb9cbd2f64178c9257c33fac607b3d48c60c