consolidate ServerTypeCheck.do_naming 2/3
Summary:
I continue my quest to clean up ServerTypeCheck.do_naming, in this diff by inlining update_naming_table and declare_names.

This diff also consolidates cgroup-profiling. Before this diff it had stage "update_deps" for updating two of the naming tables, and stage "naming" for the third. Now it just has a single cgroup for all three naming-tables.

This diff also changes the meaning of UPDATING_DEPS_END telemetry. Before it combined deps-table and fwd-table. Now it solely describes deps-table. This brings it closer in line to its use elsewhere in ServerInitCommon, where it solely times the update to the deps-table.

Differential Revision: D30883943

fbshipit-source-id: 53d5dd1fb8d14559ac12dd91f49291579700f64a