show all hack logs upon integration test failure
Summary:
When we run integration tests, and they fail, I really want to see the server+monitor logs associated with that failure. I want to be able to debug with Hh_logger.log.

That's what this diff provides. It gathers the .old log as well as the current log (in case the server died and got restarted by the test), and it prints them.

Differential Revision: D30709082

fbshipit-source-id: 247f2ae440905f62a326be3154200921d6e0c924