introduce Naming_types.name_kind
Summary:
Currently, our naming-sqlite symbols table stores a "FLAGS" field with the value 0 for classes, 1 for typedefs, 2 for recorddefs, and 0 for funs/consts.

It feels weird that we have the same value "0" for both classes and funs and consts.

Notionally in the future we shouldn't need to store FLAGS at all. We should merely fetch the decl for a particular name, and get back the appropriate decl, and check its type.

But we're not in that future yet. Current downstream consumers of naming-sqlite need to know what kind of name they're getting back even before they can fetch decls. They need this in two places:
1. Naming_heap has three different "blocked" heaps to store an in-memory-delta of symbols that have been deleted since the naming-sqlite, one for blocked funs, one for blocked consts, one for blocked types. If we merely look up a dephash in naming-sqlite, we don't know which of the three "blocked" tables to consult. One future option might be to consolidate into just a single blocked heap. But for now, I'm not going to do that.
2. Naming_heap caches the reads it does into either type, fun or const heap. If it caches the reads into the type heap, it additionally stores whether it was class/typedef/recorddef. Unless naming-sqlite get_path_by_dephash reports back the kind and subkind, then reads done by get_path_by_dephash won't be able to cache in Naming_heap. This isn't the end of the world but it would make things more complicated.

THEREFORE, I am changing the meaning of naming-sqlite FLAGS column. I will switch it to store "3" for funs and "4" for consts. We can build upon this:
1. the current diff uses it to clean up Naming_sqlite.insertion_error
2. the next diff in the stack uses it to consolidate Naming_sqlite.(FunsTable,TypesTable,ConstTable) with all their duplicate code, into just Naming_sqlite.SymbolsTable
3. Future diffs beyond this will use it to implement Naming_heap.get_path_by_dephash in a way that respects the blocked-heaps, and will use that to power dephash->filename lookup by the reverse naming table rather than Typing_deps.Files.ifiles.

This diff looks huge because the sqlite format has to be changed for both ocaml and rust. But it's really kind of small.

**Guidance for reviewer**

1. Start with naming_types.{mli,ml}
2. Next read naming_sqlite.{mli,ml}
3. The rest of the changes are just rust implementations of the same...
4. Read oxidized/manual/naming_types_impl.rs
5. Read name_index/sqlite/datatype.rs, names.rs, symbols.rs

Reviewed By: bobrenjc93

Differential Revision: D30847574

fbshipit-source-id: e102bf05d7231e4e4f6b5bf15f2221079268c200